refactor : table, order 패키지 통합
