+++ src/main/java/kitchenpos/order/dto/request/ChangeOrderTableNumberOfGuestRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.table.dto.request;
+package kitchenpos.order.dto.request;
 
 public class ChangeOrderTableNumberOfGuestRequest {
