+++ src/main/java/kitchenpos/order/repository/TableGroupRepository.java
@@ -1,11 +1,11 @@
-package kitchenpos.table.repository;
+package kitchenpos.order.repository;
 
 import java.util.List;
 import java.util.Optional;
 
 import org.springframework.data.repository.Repository;
 
-import kitchenpos.table.domain.TableGroup;
+import kitchenpos.order.domain.TableGroup;
 
 public interface TableGroupRepository extends Repository<TableGroup, Long> {
