+++ src/test/java/kitchenpos/order/domain/OrderStatusTest.java
@@ -6,8 +6,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
-import kitchenpos.order.domain.OrderStatus;
-
 class OrderStatusTest {
 
     @ParameterizedTest
